+++ libs/ionic-react/src/migrations/update-3-0-1/upgrade-ionic-3-0-1.ts
@@ -1,0 +1,18 @@
+import { stripIndents } from '@angular-devkit/core/src/utils/literals';
+import {
+  chain,
+  Rule,
+  SchematicContext,
+  Tree,
+} from '@angular-devkit/schematics';
+import { readJsonInTree, updatePackagesInPackageJson } from '@nrwl/workspace';
+import * as path from 'path';
+
+export default function update(): Rule {
+  return chain([
+    updatePackagesInPackageJson(
+      path.join(__dirname, '../../../', 'migrations.json'),
+      '3.0.1'
+    ),
+  ]);
+}