+++ libs/ionic-react/src/utils/versions.ts
@@ -1,7 +1,7 @@
 export const nxtendVersion = '*';
 
-export const ionicReactVersion = '5.2.2';
-export const ionicReactRouterVersion = '5.2.2';
+export const ionicReactVersion = '5.2.3';
+export const ionicReactRouterVersion = '5.2.3';
 export const ioniconsVersion = '5.0.1';
 
 export const testingLibraryJestDomVersion = '5.11.0';